+++ src/robin_coordsmap.hpp
@@ -82,8 +82,13 @@ public:
   CoordsMap(int ncols_, const set<int> &batch_indices);
 
   // Initializations
-  pair<vector<int>, set<int>> initialize(const int *p_coords_, int nrows_,
-                                         int ncols_, bool force_remap = false);
+  vector<int> initialize(const int *p_coords_, const int nrows_,
+                         const int ncols_, const bool force_remap = false);
+
+  pair<vector<int>, set<int>> initialize_batch(const int *p_coords_,
+                                               const int nrows_,
+                                               const int ncols_,
+                                               const bool force_remap = false);
 
   // Generate strided version of the input coordinate map.
   // returns mapping: out_coord row index to in_coord row index